Alter parents to not use prototypes
The use of prototype for the parent caused repeated validates on the same schemata
object, but with different passed objects to return the parent of the first passed
object. The validate function has been changed to set the parent, and then begin
the recursive validation that it originally did, but with this parent passed to it.