Fix a bug casting dynamic subschemas
The `getType(type, obj)` function expects an entire object
as the second argument. Inside the `castProperty()` function
however, it was only getting a single property value.

A failing test is included, now which passes. Two other tests
relied on this broken functionality, so they were fixed too.

I thought the cleanest thing to do would be change the method
signature of castProperty to `fn(type, key, obj)`, but although
that means less repetition and a simpler signature it would break
compatibility. Since `castProperty()` is a public method (and I
know is used with the signature `fn(type, value)` it's best to
leave that alone.